Refactoring bundle structure.
