+++ blended-persistence-neo4j/pom.xml
@@ -27,11 +27,11 @@
   </parent>
 
   <name>${project.artifactId}</name>
-  <artifactId>blended-persistence</artifactId>
+  <artifactId>blended-persistence-neo4j</artifactId>
   <packaging>bundle</packaging>
   
   <description>
-    Provide a technology agnostic persistence API with pluggable Data Objects defined in other bundles.
+    Implement a persistence backend with Neo4J.
   </description>
 
   <properties>
@@ -48,15 +48,9 @@
     </dependency>
 
     <dependency>
-      <groupId>org.helgoboss</groupId>
-      <artifactId>domino</artifactId>
-      <version>${domino.version}</version>
-    </dependency>
-    
-    <dependency>
-      <groupId>org.scala-lang</groupId>
-      <artifactId>scala-library</artifactId>
-      <version>${scala.version}.${scala.micro.version}</version>
+      <groupId>de.wayofquality.blended</groupId>
+      <artifactId>blended-persistence</artifactId>
+      <version>${project.version}</version>
     </dependency>
 
     <dependency>
@@ -66,6 +60,14 @@
     </dependency>
 
     <!-- Test dependencies -->
+
+    <dependency>
+      <groupId>org.neo4j</groupId>
+      <artifactId>neo4j-kernel</artifactId>
+      <version>${neo4j.version}</version>
+      <type>test-jar</type>
+      <scope>test</scope>
+    </dependency>
 
     <dependency>
       <groupId>org.scalatest</groupId>
