+++ pom.xml
@@ -173,6 +173,8 @@
 
         <module>blended-neo4j-api</module>
         <module>blended-persistence</module>
+        <module>blended-persistence-neo4j</module>
+        <module>blended-persistence-h2</module>
 
         <module>blended-jolokia</module>
         <module>blended-itestsupport</module>
